Modified some api and update slidebar

+++ src/renderer/store/modules/Home.js
@@ -1,11 +1,14 @@
 import axios from 'axios';
 import api from '../../assets/config/api'
+import router from '../../router/router';
 const state = {
   balanceTotal: 0,
   balanceAddress: '',
   channels: [],
+  channelBind: {},
   revenue: 0,
-  loginStatus: -1
+  loginStatus: -1,
+  initChannelProgress: 0
 }
 // Confirm login status
 
@@ -13,8 +16,22 @@ const state = {
 const mutations = {
   'SET_BALANCE_TOTAL'(state, result) {
     state.balanceTotal = result.BalanceFormat;
-    state.balanceAddress = result.Channels[0].Address;
+    if (result.Channels.length > 0) {
+      state.balanceAddress = result.Channels[0].Address;
+    }
     state.channels = result.Channels;
+    this.dispatch('setChannelBind', localStorage.getItem('channelBindId') || '')
+  },
+  'SET_CHANNEL_BIND'(state, Id) {
+    state.channelBind = {};
+    state.channels.map(channel => {
+      if (channel.ChannelId.toString() === Id) {
+        state.channelBind = channel;
+      }
+    })
+  },
+  'SET_CHANNEL_PROGRESS'(state, progress) {
+    state.initChannelProgress = progress;
   },
   'SET_REVENUE'(state, result) {
     state.revenue = result;
@@ -26,7 +43,8 @@ const mutations = {
 const timer = {
   COUNT_INTERVAL: 5000,
   channelBalanceTotalTimer: null,
-  revenueTimer: null
+  revenueTimer: null,
+  channelInitProgress: null
 };
 const actions = {
   setChannelBalanceTotal({
@@ -52,29 +70,71 @@ const actions = {
   }) {
     axios
       .get(api.account)
-      .then(res => {
+      .then(async (res) => {
         const data = res.data;
-        if (data.Error === 0) {
-          if (data.Desc === "SUCCESS" && data.Result.Address) {
-            const result = data.Result;
-            for (let key in result) {
-              window.localStorage.setItem(key, result[key]);
+        if (data.Error === 0) { // response data
+          if (data.Desc === "SUCCESS" && data.Result.Address) { // Wallet(Account) exist
+            try {
+              const progress = await axios.get(api.channelInitProgress)
+              if (progress.data.Error === 0 && (progress.data.Result.Progress < 1)) { // but no Channel
+                rebackToCreateAccount(commit,progress.data.Result.Progress);
+                this.dispatch('getChannelInitProgress'); // Loop loading progress
+              } else if (progress.data.Error === 0) { // both Wallet and Channel exist
+                const result = data.Result;
+                for (let key in result) {
+                  window.localStorage.setItem(key, result[key]);
+                }
+                commit('SET_CURRENT_ACCOUNT', 1) // login success
+                this.dispatch("setBalanceLists"); // getBalance
+                this.dispatch("setChannelBalanceTotal"); // getAllChannels
+                this.dispatch("setRevenue");
+              }
+
+            } catch (error) {
+              console.error(error)
+              commit('SET_CURRENT_ACCOUNT', 0) // login fail
             }
-            commit('SET_CURRENT_ACCOUNT', 1) // login success
-            this.dispatch("setBalanceLists"); // getBalance
-            this.dispatch("setChannelBalanceTotal"); // getAllChannels
-            this.dispatch("setRevenue");
           } else {
             commit('SET_CURRENT_ACCOUNT', 0) // login fail
             window.localStorage.clear(); // remove all local infomation
           }
+        } else {
+          commit('SET_CURRENT_ACCOUNT', 0) // login fail Or no user
         }
       })
-      .catch(err => {
+      .catch(err => { // network wrong
         commit('SET_CURRENT_ACCOUNT', 0) // login fail
         console.error(err);
       });
+  },
+  getChannelInitProgress({commit}) {
+    clearInterval(timer.channelInitProgress);
+    timer.channelInitProgress = setInterval(() => {
+      axios.get(api.channelInitProgress).then(res => {
+        if (res.data.Error === 0) {
+          if (res.data.Result.Progress >= 1) {
+            commit('SET_CHANNEL_PROGRESS', 1)
+            clearInterval(timer.channelInitProgress);
+            window.location.href = location.origin + location.pathname; // success login link to home page
+          }
+          commit('SET_CHANNEL_PROGRESS', res.data.Result.Progress)
+        }
+      });
+    }, timer.COUNT_INTERVAL);
+  },
+  setChannelBind({
+    commit
+  }, Id) {
+    commit('SET_CHANNEL_BIND', Id)
   }
+}
+
+function rebackToCreateAccount(commit,progress) {
+  commit('SET_CURRENT_ACCOUNT', 0) // login fail
+  commit('SET_CHANNEL_PROGRESS', progress)
+  router.replace({
+    name: 'CreateAccount'
+  })
 }
 
 function requestChannelBalanceTotal(commit) {
@@ -91,7 +151,7 @@ function requestChannelBalanceTotal(commit) {
 }
 
 function requestRevenue(commit) {
-  axios.get(api.revenue).then(res => {    
+  axios.get(api.revenue).then(res => {
     if (res.data.Error === 0) {
       commit('SET_REVENUE', res.data.Result.Revenue);
     }
